jimt/dance-ai-modal-refactoring - broke up utils.ts file; move types to ai/types
