+++ apps/src/dance/types.ts
@@ -1,5 +1,4 @@
 import {LevelProperties, ProjectSources} from '../lab2/types';
-import {DANCE_AI_SOUNDS} from './ai/constants';
 
 export type SongData = {
   [key: string]: {
@@ -36,34 +35,3 @@ export interface DanceLevelProperties extends LevelProperties {
   defaultSong?: string;
   useRestrictedSongs?: boolean;
 }
-
-export enum AiOutput {
-  AI_BLOCK = 'ai_block',
-  GENERATED_BLOCKS = 'generated_blocks',
-  BOTH = 'both',
-}
-
-export type LabelMaps = {
-  [key in FieldKey]: {[id: string]: string};
-};
-
-export enum FieldKey {
-  BACKGROUND_EFFECT = 'backgroundEffect',
-  FOREGROUND_EFFECT = 'foregroundEffect',
-  BACKGROUND_PALETTE = 'backgroundColor',
-}
-
-export type GeneratedEffect = {[key in FieldKey]: string};
-
-export type GeneratedEffectScores = number[];
-
-export type MinMax = {
-  minIndividualScore: number;
-  maxTotalScore: number;
-};
-
-export interface AiFieldValue extends GeneratedEffect {
-  inputs: string[];
-}
-
-export type DanceAiSound = (typeof DANCE_AI_SOUNDS)[number];
